feat: enhance game statistics with detailed player and match data
- Introduced new types for Player, PlayerMatch, OriginalMatch, and SharedMatch to better structure player and match information.
- Implemented matchesAggregated function to combine original and shared match data, including winners and player statistics.
- Added playerAndRolesAggregated function to aggregate player statistics by roles and match outcomes.
- Enhanced existing data structures to include detailed scoring, placements, and role statistics.
- Improved handling of player roles and combinations for more comprehensive analytics.
