+++ packages/api/src/routers/game.ts
@@ -4,11 +4,7 @@ import { compareDesc } from "date-fns";
 import { and, eq, inArray, isNull, sql } from "drizzle-orm";
 import z from "zod/v4";
 
-import type {
-  roundTypes,
-  scoreSheetRoundsScore,
-  scoreSheetWinConditions,
-} from "@board-games/db/constants";
+import type { scoreSheetWinConditions } from "@board-games/db/constants";
 import type {
   insertMatchPlayerSchema,
   insertMatchSchema,
@@ -38,16 +34,16 @@ import {
   insertScoreSheetSchema,
   selectGameSchema,
 } from "@board-games/db/zodSchema";
-import {
-  baseRoundSchema,
-  combinations,
-  editScoresheetSchema,
-} from "@board-games/shared";
+import { baseRoundSchema, editScoresheetSchema } from "@board-games/shared";
 
 import analyticsServerClient from "../analytics";
 import { createTRPCRouter, protectedUserProcedure } from "../trpc";
 import { utapi } from "../uploadthing";
-import { headToHeadStats, updateRoundStatistics } from "../utils/gameStats";
+import {
+  headToHeadStats,
+  matchesAggregated,
+  playerAndRolesAggregated,
+} from "../utils/gameStats";
 
 export const gameRouter = createTRPCRouter({
   create: protectedUserProcedure
@@ -797,341 +793,11 @@ export const gameRouter = createTRPCRouter({
         },
       });
       if (!result) return null;
-      const roleStats = result.roles.reduce(
-        (acc, role) => {
-          acc[role.id] = {
-            roleId: role.id,
-            name: role.name,
-            description: role.description,
-            playerCount: 0,
-            matchCount: 0,
-            winRate: 0,
-            wins: 0,
-            losses: 0,
-            placements: {},
-            players: {},
-          };
-          return acc;
-        },
-        {} as Record<
-          number,
-          {
-            roleId: number;
-            name: string;
-            description: string | null;
-            playerCount: number;
-            matchCount: number;
-            winRate: number;
-            wins: number;
-            losses: number;
-            placements: Record<number, number>;
-            players: Record<
-              number,
-              {
-                id: number;
-                name: string;
-                isUser: boolean;
-                image: {
-                  name: string;
-                  url: string | null;
-                  type: "file" | "svg";
-                  usageType: "player" | "match" | "game";
-                } | null;
-                totalMatches: number;
-                totalWins: number;
-                totalLosses: number;
-                winRate: number;
-                placements: Record<number, number>;
-              }
-            >;
-          }
-        >,
-      );
-      const comboRoles: Record<
-        string,
-        {
-          roles: {
-            id: number;
-            name: string;
-            description: string | null;
-          }[];
-          matchCount: number;
-          wins: number;
-          losses: number;
-          players: number;
-          winRate: number;
-          placements: Record<number, number>;
-        }
-      > = {};
-      const matches: {
-        type: "original" | "shared";
-        id: number;
-        gameId: number;
-        date: Date;
-        location: {
-          type: "shared" | "linked" | "original";
-          name: string;
-        } | null;
-        won: boolean;
-        placement: number | null;
-        score: number | null;
-        name: string;
-        duration: number;
-        finished: boolean;
-        comment: string | null;
-        scoresheet: {
-          id: number;
-          parentId: number | null;
-          winCondition: (typeof scoreSheetWinConditions)[number];
-          roundScore: (typeof scoreSheetRoundsScore)[number];
-          targetScore: z.infer<typeof selectScoreSheetSchema>["targetScore"];
-          isCoop: z.infer<typeof selectScoreSheetSchema>["isCoop"];
-          rounds: {
-            id: number;
-            parentId: number | null;
-            name: string;
-            type: (typeof roundTypes)[number];
-            score: number;
-            order: number;
-          }[];
-        };
-        players: {
-          id: number;
-          type: "original" | "shared";
-          name: string;
-          isWinner: boolean | null;
-          isUser: boolean;
-          score: number | null;
-          placement: number;
-          image: {
-            name: string;
-            url: string | null;
-            type: "file" | "svg";
-            usageType: "player" | "match" | "game";
-          } | null;
-          team: {
-            id: number;
-            name: string;
-            matchId: number;
-            details: string | null;
-            createdAt: Date;
-            updatedAt: Date | null;
-          } | null;
-          playerRounds: {
-            id: number;
-            roundId: number;
-            score: number | null;
-          }[];
-          roles: {
-            id: number;
-            name: string;
-            description: string | null;
-          }[];
-        }[];
-        winners: {
-          id: number;
-          name: string;
-          isWinner: boolean | null;
-          score: number | null;
-          team: {
-            id: number;
-            name: string;
-            matchId: number;
-            details: string | null;
-            createdAt: Date;
-            updatedAt: Date | null;
-          } | null;
-        }[];
-      }[] = result.matches.map((match) => {
-        const winners = match.matchPlayers.filter((player) => player.winner);
-        const foundPlayer = match.matchPlayers.find((p) => p.player.isUser);
 
-        return {
-          type: "original" as const,
-          id: match.id,
-          gameId: match.gameId,
-          date: match.date,
-          location: match.location
-            ? {
-                type: "original" as const,
-                name: match.location.name,
-              }
-            : null,
-          won: match.finished ? (foundPlayer?.winner ?? false) : false,
-          placement: match.finished ? (foundPlayer?.placement ?? null) : null,
-          score: match.finished ? (foundPlayer?.score ?? null) : null,
-          name: match.name,
-          comment: match.comment,
-          duration: match.duration,
-          finished: match.finished,
-          scoresheet: {
-            id: match.scoresheet.id,
-            parentId: match.scoresheet.parentId,
-            winCondition: match.scoresheet.winCondition,
-            roundScore: match.scoresheet.roundsScore,
-            targetScore: match.scoresheet.targetScore,
-            isCoop: match.scoresheet.isCoop,
-            rounds: match.scoresheet.rounds.map((round) => ({
-              id: round.id,
-              parentId: round.parentId,
-              name: round.name,
-              type: round.type,
-              score: round.score,
-              order: round.order,
-            })),
-          },
-          players: match.matchPlayers.map((player) => {
-            return {
-              id: player.player.id,
-              type: "original" as const,
-              name: player.player.name,
-              isWinner: player.winner,
-              isUser: player.player.isUser,
-              score: player.score,
-              image: player.player.image,
-              team: player.team,
-              placement: player.placement ?? 0,
-              playerRounds: player.playerRounds.map((round) => ({
-                id: round.id,
-                roundId: round.roundId,
-                score: round.score,
-              })),
-              roles: player.roles.map((role) => ({
-                id: role.id,
-                name: role.name,
-                description: role.description,
-              })),
-            };
-          }),
-          winners: winners.map((player) => {
-            return {
-              id: player.player.id,
-              name: player.player.name,
-              isWinner: player.winner,
-              score: player.score,
-              team: player.team,
-            };
-          }),
-        };
-      });
-      for (const returnedShareMatch of result.sharedGameMatches) {
-        const winners = returnedShareMatch.sharedMatchPlayers.filter(
-          (returnedSharedMatchPlayer) =>
-            returnedSharedMatchPlayer.matchPlayer.winner,
-        );
-        const foundSharedPlayer = returnedShareMatch.sharedMatchPlayers.find(
-          (p) => p.sharedPlayer?.linkedPlayer?.isUser,
-        )?.matchPlayer;
-        const mSharedLocation = returnedShareMatch.sharedLocation;
-        const mLinkedLocation = mSharedLocation?.linkedLocation;
-        const mappedShareMatch = {
-          type: "shared" as const,
-          id: returnedShareMatch.id,
-          gameId: returnedShareMatch.sharedGameId,
-          name: returnedShareMatch.match.name,
-          comment: returnedShareMatch.match.comment,
-          date: returnedShareMatch.match.date,
-          location: mSharedLocation
-            ? {
-                type: mLinkedLocation
-                  ? ("linked" as const)
-                  : ("shared" as const),
-                name: mLinkedLocation?.name ?? mSharedLocation.location.name,
-              }
-            : null,
-          duration: returnedShareMatch.match.duration,
-          finished: returnedShareMatch.match.finished,
-          scoresheet: {
-            id: returnedShareMatch.match.scoresheet.id,
-            parentId: returnedShareMatch.match.scoresheet.parentId,
-            winCondition: returnedShareMatch.match.scoresheet.winCondition,
-            roundScore: returnedShareMatch.match.scoresheet.roundsScore,
-            targetScore: returnedShareMatch.match.scoresheet.targetScore,
-            isCoop: returnedShareMatch.match.scoresheet.isCoop,
-            rounds: returnedShareMatch.match.scoresheet.rounds.map((round) => ({
-              id: round.id,
-              parentId: round.parentId,
-              name: round.name,
-              type: round.type,
-              score: round.score,
-              order: round.order,
-            })),
-          },
-          won: returnedShareMatch.match.finished
-            ? (foundSharedPlayer?.winner ?? false)
-            : false,
-          placement: returnedShareMatch.match.finished
-            ? (foundSharedPlayer?.placement ?? null)
-            : null,
-          score: returnedShareMatch.match.finished
-            ? (foundSharedPlayer?.score ?? null)
-            : null,
-          players: returnedShareMatch.sharedMatchPlayers
-            .map((returnedSharedMatchPlayer) => {
-              if (returnedSharedMatchPlayer.sharedPlayer === null) return null;
-              const linkedPlayer =
-                returnedSharedMatchPlayer.sharedPlayer.linkedPlayer;
-              return {
-                type:
-                  linkedPlayer !== null
-                    ? ("original" as const)
-                    : ("shared" as const),
-                id:
-                  linkedPlayer !== null
-                    ? linkedPlayer.id
-                    : returnedSharedMatchPlayer.sharedPlayer.playerId,
-                name:
-                  linkedPlayer !== null
-                    ? linkedPlayer.name
-                    : returnedSharedMatchPlayer.sharedPlayer.player.name,
-                isUser: linkedPlayer?.isUser ?? false,
-                isWinner: returnedSharedMatchPlayer.matchPlayer.winner,
-                score: returnedSharedMatchPlayer.matchPlayer.score,
-                placement: returnedSharedMatchPlayer.matchPlayer.placement ?? 0,
-                team: returnedSharedMatchPlayer.matchPlayer.team,
-                image:
-                  linkedPlayer !== null
-                    ? linkedPlayer.image
-                    : returnedSharedMatchPlayer.sharedPlayer.player.image,
-                playerRounds:
-                  returnedSharedMatchPlayer.matchPlayer.playerRounds.map(
-                    (round) => ({
-                      id: round.id,
-                      roundId: round.roundId,
-                      score: round.score,
-                    }),
-                  ),
-                roles: returnedSharedMatchPlayer.matchPlayer.roles.map(
-                  (role) => ({
-                    id: role.id,
-                    name: role.name,
-                    description: role.description,
-                  }),
-                ),
-              };
-            })
-            .filter((player) => player !== null),
-          winners: winners
-            .map((returnedSharedMatchPlayer) => {
-              if (returnedSharedMatchPlayer.sharedPlayer === null) return null;
-              const linkedPlayer =
-                returnedSharedMatchPlayer.sharedPlayer.linkedPlayer;
-              return {
-                type: "shared" as const,
-                id: returnedSharedMatchPlayer.sharedPlayer.playerId,
-                name:
-                  linkedPlayer !== null
-                    ? linkedPlayer.name
-                    : returnedSharedMatchPlayer.sharedPlayer.player.name,
-                isWinner: returnedSharedMatchPlayer.matchPlayer.winner,
-                score: returnedSharedMatchPlayer.matchPlayer.score,
-                team: returnedSharedMatchPlayer.matchPlayer.team,
-              };
-            })
-            .filter((winner) => winner !== null),
-        };
-        matches.push(mappedShareMatch);
-      }
+      const matches = matchesAggregated(
+        result.matches,
+        result.sharedGameMatches,
+      );
       matches.sort((a, b) => compareDesc(a.date, b.date));
       const gameScoresheets: {
         id: number;
@@ -1191,828 +857,8 @@ export const gameRouter = createTRPCRouter({
           }
         });
       });
-      const seenRoleMatchIds = new Map<number, Set<number>>();
-      const seenRoleComboMatchIds = new Map<string, Set<number>>();
-      const roleBasedPlayerSeenMatchIds = new Map<
-        number,
-        Map<string, Set<number>>
-      >();
-      const playerSeenRoleComboMatchIds = new Map<
-        string,
-        Map<string, Set<number>>
-      >();
-      const playerSeenRoleMatchIds = new Map<
-        string,
-        Map<number, Set<number>>
-      >();
-      const players = matches.reduce(
-        (acc, match) => {
-          if (!match.finished) return acc;
-          const currentScoresheet = match.scoresheet;
-          const isCoop = currentScoresheet.isCoop;
-          match.players.forEach((player) => {
-            const accPlayer = acc[`${player.type}-${player.id}`];
-            player.roles.forEach((role) => {
-              const accRole = roleStats[role.id];
-              if (!accRole) {
-                roleStats[role.id] = {
-                  roleId: role.id,
-                  name: role.name,
-                  description: role.description,
-                  playerCount: 1,
-                  placements:
-                    player.placement > 0 ? { [player.placement]: 1 } : {},
-                  wins: player.isWinner ? 1 : 0,
-                  losses: player.isWinner ? 0 : 1,
-                  matchCount: 1,
-                  winRate: player.isWinner ? 1 : 0,
-                  players: {
-                    [player.id]: {
-                      id: player.id,
-                      name: player.name,
-                      isUser: player.isUser,
-                      image: player.image,
-                      totalMatches: 1,
-                      totalWins: player.isWinner ? 1 : 0,
-                      totalLosses: player.isWinner ? 0 : 1,
-                      winRate: player.isWinner ? 1 : 0,
-                      placements:
-                        player.placement > 0 ? { [player.placement]: 1 } : {},
-                    },
-                  },
-                };
-                seenRoleMatchIds.set(role.id, new Set());
-                let roleBasedPlayerMap = roleBasedPlayerSeenMatchIds.get(
-                  role.id,
-                );
-                if (!roleBasedPlayerMap) {
-                  roleBasedPlayerMap = new Map();
-                  const playerSet = new Set<number>();
-                  playerSet.add(match.id);
-                  roleBasedPlayerMap.set(
-                    `${player.type}-${player.id}`,
-                    playerSet,
-                  );
-                  roleBasedPlayerSeenMatchIds.set(role.id, roleBasedPlayerMap);
-                } else {
-                  let playerSet = roleBasedPlayerMap.get(
-                    `${player.type}-${player.id}`,
-                  );
-                  if (!playerSet) {
-                    playerSet = new Set();
-                    playerSet.add(match.id);
-                    roleBasedPlayerMap.set(
-                      `${player.type}-${player.id}`,
-                      playerSet,
-                    );
-                  } else {
-                    playerSet.add(match.id);
-                  }
-                }
-              } else {
-                accRole.playerCount++;
-                if (player.placement > 0) {
-                  accRole.placements[player.placement] =
-                    (accRole.placements[player.placement] ?? 0) + 1;
-                }
-                accRole.wins += player.isWinner ? 1 : 0;
-                accRole.losses += player.isWinner ? 0 : 1;
-                let roleMatchSet = seenRoleMatchIds.get(role.id);
-                if (!roleMatchSet) {
-                  roleMatchSet = new Set();
-                  roleMatchSet.add(match.id);
-                  seenRoleMatchIds.set(role.id, roleMatchSet);
-                  accRole.matchCount++;
-                } else {
-                  if (!roleMatchSet.has(match.id)) {
-                    roleMatchSet.add(match.id);
-                    accRole.matchCount++;
-                  }
-                }
-
-                accRole.winRate =
-                  accRole.wins / (accRole.wins + accRole.losses);
-                const accPlayer = accRole.players[player.id];
-                if (!accPlayer) {
-                  accRole.players[player.id] = {
-                    id: player.id,
-                    name: player.name,
-                    isUser: player.isUser,
-                    image: player.image,
-                    totalMatches: 1,
-                    totalWins: player.isWinner ? 1 : 0,
-                    totalLosses: player.isWinner ? 0 : 1,
-                    winRate: player.isWinner ? 1 : 0,
-                    placements:
-                      player.placement > 0 ? { [player.placement]: 1 } : {},
-                  };
-                  let roleBasedPlayerMatchIds = roleBasedPlayerSeenMatchIds.get(
-                    role.id,
-                  );
-                  if (!roleBasedPlayerMatchIds) {
-                    roleBasedPlayerMatchIds = new Map();
-                    const playerSet = new Set<number>();
-                    playerSet.add(match.id);
-                    roleBasedPlayerMatchIds.set(
-                      `${player.type}-${player.id}`,
-                      playerSet,
-                    );
-                    roleBasedPlayerSeenMatchIds.set(
-                      role.id,
-                      roleBasedPlayerMatchIds,
-                    );
-                  } else {
-                    let playerSet = roleBasedPlayerMatchIds.get(
-                      `${player.type}-${player.id}`,
-                    );
-                    if (!playerSet) {
-                      playerSet = new Set();
-                      playerSet.add(match.id);
-                      roleBasedPlayerMatchIds.set(
-                        `${player.type}-${player.id}`,
-                        playerSet,
-                      );
-                    } else {
-                      playerSet.add(match.id);
-                    }
-                  }
-                } else {
-                  let roleBasedPlayerMatchIds = roleBasedPlayerSeenMatchIds.get(
-                    role.id,
-                  );
-                  if (!roleBasedPlayerMatchIds) {
-                    roleBasedPlayerMatchIds = new Map();
-                    const playerSet = new Set<number>();
-                    playerSet.add(match.id);
-                    roleBasedPlayerMatchIds.set(
-                      `${player.type}-${player.id}`,
-                      playerSet,
-                    );
-                    roleBasedPlayerSeenMatchIds.set(
-                      role.id,
-                      roleBasedPlayerMatchIds,
-                    );
-                    accPlayer.totalMatches++;
-                  } else {
-                    const playerSet = roleBasedPlayerMatchIds.get(
-                      `${player.type}-${player.id}`,
-                    );
-                    if (!playerSet) {
-                      const playerSet = new Set<number>();
-                      playerSet.add(match.id);
-                      roleBasedPlayerMatchIds.set(
-                        `${player.type}-${player.id}`,
-                        playerSet,
-                      );
-                      accPlayer.totalMatches++;
-                    } else if (!playerSet.has(match.id)) {
-                      playerSet.add(match.id);
-                      accPlayer.totalMatches++;
-                    }
-                  }
-                  accPlayer.totalWins += player.isWinner ? 1 : 0;
-                  accPlayer.winRate =
-                    accPlayer.totalWins /
-                    (accPlayer.totalWins + accPlayer.totalLosses);
-                  if (player.placement > 0) {
-                    accPlayer.placements[player.placement] =
-                      (accPlayer.placements[player.placement] ?? 0) + 1;
-                  }
-                }
-              }
-            });
-            if (!accPlayer) {
-              const tempPlacements: Record<number, number> = {};
-              const tempPlayerCount: Record<
-                number,
-                {
-                  playerCount: number;
-                  placements: Record<number, number>;
-                  wins: number;
-                  plays: number;
-                }
-              > = {};
-              const tempScoresheets: Record<
-                number,
-                {
-                  id: number;
-                  bestScore: number | null;
-                  worstScore: number | null;
-                  scores: {
-                    date: Date;
-                    score: number | null;
-                    isWin: boolean;
-                  }[];
-                  winRate: number;
-                  plays: number;
-                  wins: number;
-                  placements: Record<number, number>;
-                  rounds: Record<
-                    number,
-                    {
-                      id: number;
-                      bestScore: number | null;
-                      worstScore: number | null;
-                      scores: {
-                        date: Date;
-                        score: number | null;
-                        isWin: boolean;
-                      }[];
-                    }
-                  >;
-                }
-              > = {};
-
-              if (!isCoop) {
-                tempPlacements[player.placement] = 1;
-                tempPlayerCount[match.players.length] = {
-                  playerCount: match.players.length,
-                  placements: {
-                    [player.placement]: 1,
-                  },
-                  wins: player.isWinner ? 1 : 0,
-                  plays: 1,
-                };
-              }
-              if (currentScoresheet.parentId) {
-                const tempPlayerRounds = updateRoundStatistics(
-                  player.playerRounds,
-                  currentScoresheet.rounds,
-                  currentScoresheet.winCondition,
-                  match.date,
-                  player.isWinner ?? false,
-                );
-                tempScoresheets[currentScoresheet.parentId] = {
-                  id: currentScoresheet.parentId,
-                  bestScore: player.score,
-                  worstScore: player.score,
-                  scores: [
-                    {
-                      date: match.date,
-                      score: player.score,
-                      isWin: player.isWinner ?? false,
-                    },
-                  ],
-                  winRate: player.isWinner ? 1 : 0,
-                  plays: 1,
-                  wins: player.isWinner ? 1 : 0,
-                  placements: !isCoop ? tempPlacements : {},
-                  rounds: tempPlayerRounds,
-                };
-              }
-              let playerSeenRoleMatch = playerSeenRoleMatchIds.get(
-                `${player.type}-${player.id}`,
-              );
-              if (!playerSeenRoleMatch) {
-                playerSeenRoleMatchIds.set(
-                  `${player.type}-${player.id}`,
-                  new Map(),
-                );
-                playerSeenRoleMatch = playerSeenRoleMatchIds.get(
-                  `${player.type}-${player.id}`,
-                );
-              }
-              acc[`${player.type}-${player.id}`] = {
-                id: player.id,
-                type: player.type,
-                name: player.name,
-                isUser: player.isUser,
-                coopWins: player.isWinner && isCoop ? 1 : 0,
-                competitiveWins: player.isWinner && !isCoop ? 1 : 0,
-                coopWinRate: player.isWinner && isCoop ? 1 : 0,
-                competitiveWinRate: player.isWinner && !isCoop ? 1 : 0,
-                coopMatches: isCoop ? 1 : 0,
-                competitiveMatches: !isCoop ? 1 : 0,
-                coopScores: isCoop
-                  ? [
-                      {
-                        date: match.date,
-                        score: player.score,
-                        isWin: player.isWinner ?? false,
-                      },
-                    ]
-                  : [],
-                competitiveScores: !isCoop
-                  ? [
-                      {
-                        date: match.date,
-                        score: player.score,
-                        isWin: player.isWinner ?? false,
-                      },
-                    ]
-                  : [],
-                image: player.image,
-                placements: !isCoop ? tempPlacements : {},
-                streaks: {
-                  current: {
-                    type: player.isWinner ? "win" : "loss",
-                    count: 1,
-                  },
-                  longest: {
-                    wins: player.isWinner ? 1 : 0,
-                    losses: player.isWinner ? 0 : 1,
-                  },
-                },
-                recentForm: player.isWinner ? ["win"] : ["loss"],
-                playerCount: !isCoop ? tempPlayerCount : {},
-                scoresheets: tempScoresheets,
-                roles: player.roles.reduce(
-                  (acc, role) => {
-                    acc[role.id] = {
-                      roleId: role.id,
-                      name: role.name,
-                      description: role.description,
-                      matchCount: 1,
-                      winRate: player.isWinner ? 1 : 0,
-                      wins: player.isWinner ? 1 : 0,
-                      losses: player.isWinner ? 0 : 1,
-                      placements:
-                        player.placement > 0 ? { [player.placement]: 1 } : {},
-                    };
-                    let playerRoleMatchIds = playerSeenRoleMatchIds.get(
-                      `${player.type}-${player.id}`,
-                    );
-                    if (!playerRoleMatchIds) {
-                      playerSeenRoleMatchIds.set(
-                        `${player.type}-${player.id}`,
-                        new Map(),
-                      );
-                      playerRoleMatchIds = playerSeenRoleMatchIds.get(
-                        `${player.type}-${player.id}`,
-                      );
-                    }
-
-                    let playerRoleMap = playerRoleMatchIds?.get(role.id);
-                    if (!playerRoleMap) {
-                      playerRoleMap = new Set();
-                      playerRoleMap.add(match.id);
-                      playerRoleMatchIds?.set(role.id, playerRoleMap);
-                    } else {
-                      playerRoleMap.add(match.id);
-                    }
-                    return acc;
-                  },
-                  {} as Record<
-                    number,
-                    {
-                      roleId: number;
-                      name: string;
-                      description: string | null;
-                      matchCount: number;
-                      winRate: number;
-                      wins: number;
-                      losses: number;
-                      placements: Record<number, number>;
-                    }
-                  >,
-                ),
-                roleCombos: {},
-              };
-            } else {
-              accPlayer.recentForm.push(player.isWinner ? "win" : "loss");
-              const current = accPlayer.streaks.current;
-              let playerRoleMatchIds = playerSeenRoleMatchIds.get(
-                `${player.type}-${player.id}`,
-              );
-              if (!playerRoleMatchIds) {
-                playerSeenRoleMatchIds.set(
-                  `${player.type}-${player.id}`,
-                  new Map(),
-                );
-                playerRoleMatchIds = playerSeenRoleMatchIds.get(
-                  `${player.type}-${player.id}`,
-                );
-              }
-              player.roles.forEach((role) => {
-                const accRole = accPlayer.roles[role.id];
-                if (!accRole) {
-                  accPlayer.roles[role.id] = {
-                    roleId: role.id,
-                    name: role.name,
-                    description: role.description,
-                    matchCount: 1,
-                    winRate: player.isWinner ? 1 : 0,
-                    wins: player.isWinner ? 1 : 0,
-                    losses: player.isWinner ? 0 : 1,
-                    placements:
-                      player.placement > 0 ? { [player.placement]: 1 } : {},
-                  };
-
-                  let playerRoleMap = playerRoleMatchIds?.get(role.id);
-                  if (!playerRoleMap) {
-                    playerRoleMap = new Set();
-                    playerRoleMap.add(match.id);
-                    playerRoleMatchIds?.set(role.id, playerRoleMap);
-                  } else {
-                    playerRoleMap.add(match.id);
-                  }
-                } else {
-                  let playerSeenRoleMatch = playerSeenRoleMatchIds.get(
-                    `${player.type}-${player.id}`,
-                  );
-                  if (!playerSeenRoleMatch) {
-                    playerSeenRoleMatchIds.set(
-                      `${player.type}-${player.id}`,
-                      new Map(),
-                    );
-                    playerSeenRoleMatch = playerSeenRoleMatchIds.get(
-                      `${player.type}-${player.id}`,
-                    );
-                  }
-                  let playerRoleMatch = playerSeenRoleMatch?.get(role.id);
-                  if (!playerRoleMatch) {
-                    playerRoleMatch = new Set();
-                    playerRoleMatch.add(match.id);
-                    playerSeenRoleMatch?.set(role.id, playerRoleMatch);
-                    accRole.matchCount++;
-                  } else {
-                    if (!playerRoleMatch.has(match.id)) {
-                      accRole.matchCount++;
-                    }
-                  }
-
-                  accRole.winRate =
-                    (accRole.wins + (player.isWinner ? 1 : 0)) /
-                    accRole.matchCount;
-                  accRole.wins += player.isWinner ? 1 : 0;
-                  accRole.losses += player.isWinner ? 0 : 1;
-
-                  if (player.placement > 0) {
-                    accRole.placements[player.placement] =
-                      (accRole.placements[player.placement] ?? 0) + 1;
-                  }
-                }
-              });
-              if (
-                (player.isWinner && current.type === "win") ||
-                (!player.isWinner && current.type === "loss")
-              ) {
-                current.count = current.count + 1;
-              } else {
-                current.type = player.isWinner ? "win" : "loss";
-                current.count = 1;
-              }
-
-              const longest = accPlayer.streaks.longest;
-              if (current.count > longest.wins && current.type === "win") {
-                longest.wins = current.count;
-              }
-              if (current.count > longest.losses && current.type === "loss") {
-                longest.losses = current.count;
-              }
-              if (isCoop) {
-                if (player.isWinner) accPlayer.coopWins++;
-                accPlayer.coopMatches++;
-                accPlayer.coopScores.push({
-                  date: match.date,
-                  score: player.score,
-                  isWin: player.isWinner ?? false,
-                });
-              } else {
-                if (player.isWinner) accPlayer.competitiveWins++;
-                accPlayer.competitiveMatches++;
-                accPlayer.placements[player.placement] =
-                  (accPlayer.placements[player.placement] ?? 0) + 1;
-                accPlayer.competitiveScores.push({
-                  date: match.date,
-                  score: player.score,
-                  isWin: player.isWinner ?? false,
-                });
-                const playerCount = accPlayer.playerCount[match.players.length];
-                if (playerCount) {
-                  playerCount.plays = playerCount.plays + 1;
-                  playerCount.wins =
-                    playerCount.wins + (player.isWinner ? 1 : 0);
-                  playerCount.placements[player.placement] =
-                    (playerCount.placements[player.placement] ?? 0) + 1;
-                } else {
-                  accPlayer.playerCount[match.players.length] = {
-                    playerCount: match.players.length,
-                    placements: {
-                      [player.placement]: 1,
-                    },
-                    wins: player.isWinner ? 1 : 0,
-                    plays: 1,
-                  };
-                }
-              }
-              if (currentScoresheet.parentId) {
-                const accScoresheet =
-                  accPlayer.scoresheets[currentScoresheet.parentId];
-                if (!accScoresheet) {
-                  const tempPlayerRounds = updateRoundStatistics(
-                    player.playerRounds,
-                    currentScoresheet.rounds,
-                    currentScoresheet.winCondition,
-                    match.date,
-                    player.isWinner ?? false,
-                  );
-                  accPlayer.scoresheets[currentScoresheet.parentId] = {
-                    id: currentScoresheet.parentId,
-                    bestScore: player.score,
-                    worstScore: player.score,
-                    scores: [
-                      {
-                        date: match.date,
-                        score: player.score,
-                        isWin: player.isWinner ?? false,
-                      },
-                    ],
-                    winRate: player.isWinner ? 1 : 0,
-                    plays: 1,
-                    wins: player.isWinner ? 1 : 0,
-                    placements: accPlayer.placements,
-                    rounds: tempPlayerRounds,
-                  };
-                } else {
-                  accScoresheet.plays++;
-                  accScoresheet.wins += player.isWinner ? 1 : 0;
-                  if (player.score !== null) {
-                    if (currentScoresheet.winCondition === "Lowest Score") {
-                      accScoresheet.bestScore = accScoresheet.bestScore
-                        ? Math.min(accScoresheet.bestScore, player.score)
-                        : player.score;
-                      accScoresheet.worstScore = accScoresheet.worstScore
-                        ? Math.max(accScoresheet.worstScore, player.score)
-                        : player.score;
-                    } else if (
-                      currentScoresheet.winCondition === "Highest Score"
-                    ) {
-                      accScoresheet.bestScore = accScoresheet.bestScore
-                        ? Math.max(accScoresheet.bestScore, player.score)
-                        : player.score;
-                      accScoresheet.worstScore = accScoresheet.worstScore
-                        ? Math.min(accScoresheet.worstScore, player.score)
-                        : player.score;
-                    }
-                  }
-                  accScoresheet.scores.push({
-                    date: match.date,
-                    score: player.score,
-                    isWin: player.isWinner ?? false,
-                  });
-                  player.playerRounds.forEach((pRound) => {
-                    const foundRound = currentScoresheet.rounds.find(
-                      (round) => round.id === pRound.roundId,
-                    );
-                    if (foundRound?.parentId) {
-                      const accPlayerRound =
-                        accScoresheet.rounds[foundRound.parentId];
-                      if (!accPlayerRound) {
-                        accScoresheet.rounds[foundRound.parentId] = {
-                          id: foundRound.id,
-                          bestScore:
-                            currentScoresheet.winCondition === "Lowest Score" ||
-                            currentScoresheet.winCondition === "Highest Score"
-                              ? pRound.score
-                              : null,
-                          worstScore:
-                            currentScoresheet.winCondition === "Lowest Score" ||
-                            currentScoresheet.winCondition === "Highest Score"
-                              ? pRound.score
-                              : null,
-                          scores: [
-                            {
-                              date: match.date,
-                              score: pRound.score,
-                              isWin: player.isWinner ?? false,
-                            },
-                          ],
-                        };
-                      } else {
-                        if (pRound.score !== null) {
-                          if (
-                            currentScoresheet.winCondition === "Lowest Score"
-                          ) {
-                            accPlayerRound.bestScore = Math.min(
-                              accPlayerRound.bestScore ?? 0,
-                              pRound.score,
-                            );
-                            accPlayerRound.worstScore = Math.max(
-                              accPlayerRound.worstScore ?? 0,
-                              pRound.score,
-                            );
-                          } else if (
-                            currentScoresheet.winCondition === "Highest Score"
-                          ) {
-                            accPlayerRound.bestScore = Math.max(
-                              accPlayerRound.bestScore ?? 0,
-                              pRound.score,
-                            );
-                            accPlayerRound.worstScore = Math.min(
-                              accPlayerRound.worstScore ?? 0,
-                              pRound.score,
-                            );
-                          }
-                        }
-                        accPlayerRound.scores.push({
-                          date: match.date,
-                          score: pRound.score,
-                          isWin: player.isWinner ?? false,
-                        });
-                      }
-                    }
-                  });
-                  if (!isCoop)
-                    accScoresheet.placements[player.placement] =
-                      (accScoresheet.placements[player.placement] ?? 0) + 1;
-                }
-              }
-            }
-            if (player.roles.length >= 2) {
-              const accPlayer = acc[`${player.type}-${player.id}`];
-              if (!accPlayer) {
-                console.error(
-                  `Player ${player.type}-${player.id} not found in accumulator`,
-                );
-              }
-              const roleCombos = combinations(player.roles, 2);
-              for (const roleCombo of roleCombos) {
-                const sortedCombo = roleCombo
-                  .slice()
-                  .sort((a, b) => a.name.localeCompare(b.name));
-                const roleComboKey = sortedCombo.map((r) => r.name).join(" + ");
-                const globalCombo = comboRoles[roleComboKey];
-                if (!globalCombo) {
-                  comboRoles[roleComboKey] = {
-                    roles: sortedCombo.map((r) => ({
-                      id: r.id,
-                      name: r.name,
-                      description: r.description,
-                    })),
-                    matchCount: 1,
-                    players: 1,
-                    wins: player.isWinner ? 1 : 0,
-                    losses: player.isWinner ? 0 : 1,
-                    winRate: player.isWinner ? 1 : 0,
-                    placements:
-                      player.placement > 0 ? { [player.placement]: 1 } : {},
-                  };
-                  seenRoleComboMatchIds.set(roleComboKey, new Set());
-                } else {
-                  globalCombo.players++;
-                  if (!seenRoleComboMatchIds.get(roleComboKey)?.has(match.id)) {
-                    globalCombo.matchCount++;
-                  }
-                  globalCombo.wins += player.isWinner ? 1 : 0;
-                  globalCombo.losses += player.isWinner ? 0 : 1;
-                  globalCombo.winRate =
-                    globalCombo.wins / (globalCombo.wins + globalCombo.losses);
-                  if (player.placement > 0) {
-                    globalCombo.placements[player.placement] =
-                      (globalCombo.placements[player.placement] ?? 0) + 1;
-                  }
-                }
-                const playerCombo = accPlayer?.roleCombos[roleComboKey];
-                if (accPlayer && playerCombo) {
-                  if (
-                    !playerSeenRoleComboMatchIds
-                      .get(`${player.type}-${player.id}`)
-                      ?.get(roleComboKey)
-                      ?.has(match.id)
-                  ) {
-                    playerCombo.matchCount++;
-                  }
-                  playerCombo.wins += player.isWinner ? 1 : 0;
-                  playerCombo.losses += player.isWinner ? 0 : 1;
-                  playerCombo.winRate =
-                    playerCombo.wins / (playerCombo.wins + playerCombo.losses);
-                  if (player.placement > 0) {
-                    playerCombo.placements[player.placement] =
-                      (playerCombo.placements[player.placement] ?? 0) + 1;
-                  }
-                } else if (accPlayer) {
-                  accPlayer.roleCombos[roleComboKey] = {
-                    roles: sortedCombo.map((r) => ({
-                      id: r.id,
-                      name: r.name,
-                      description: r.description,
-                    })),
-                    matchCount: 1,
-                    wins: player.isWinner ? 1 : 0,
-                    losses: player.isWinner ? 0 : 1,
-                    winRate: player.isWinner ? 1 : 0,
-                    placements:
-                      player.placement > 0 ? { [player.placement]: 1 } : {},
-                  };
-                  playerSeenRoleComboMatchIds.set(
-                    `${player.type}-${player.id}`,
-                    new Map(),
-                  );
-                  playerSeenRoleComboMatchIds
-                    .get(`${player.type}-${player.id}`)
-                    ?.set(roleComboKey, new Set());
-                }
-              }
-            }
-          });
-          return acc;
-        },
-        {} as Record<
-          string,
-          {
-            id: number;
-            type: "original" | "shared";
-            name: string;
-            isUser: boolean;
-            coopWinRate: number;
-            competitiveWins: number;
-            coopWins: number;
-            competitiveWinRate: number;
-            coopMatches: number;
-            competitiveMatches: number;
-            coopScores: {
-              date: Date;
-              score: number | null;
-              isWin: boolean;
-            }[];
-            competitiveScores: {
-              date: Date;
-              score: number | null;
-              isWin: boolean;
-            }[];
-            image: {
-              name: string;
-              url: string | null;
-              type: "file" | "svg";
-              usageType: "player" | "match" | "game";
-            } | null;
-            placements: Record<number, number>;
-            streaks: {
-              current: { type: "win" | "loss"; count: number };
-              longest: { wins: number; losses: number };
-            };
-            recentForm: ("win" | "loss")[];
-            playerCount: Record<
-              number,
-              {
-                playerCount: number;
-                placements: Record<number, number>;
-                wins: number;
-                plays: number;
-              }
-            >;
-            scoresheets: Record<
-              number,
-              {
-                id: number;
-                bestScore: number | null;
-                worstScore: number | null;
-                scores: {
-                  date: Date;
-                  score: number | null;
-                  isWin: boolean;
-                }[];
-                winRate: number;
-                plays: number;
-                wins: number;
-                placements: Record<number, number>;
-                rounds: Record<
-                  number,
-                  {
-                    id: number;
-                    bestScore: number | null;
-                    worstScore: number | null;
-                    scores: {
-                      date: Date;
-                      score: number | null;
-                      isWin: boolean;
-                    }[];
-                  }
-                >;
-              }
-            >;
-            roles: Record<
-              number,
-              {
-                roleId: number;
-                name: string;
-                description: string | null;
-                matchCount: number;
-                winRate: number;
-                wins: number;
-                losses: number;
-                placements: Record<number, number>;
-              }
-            >;
-            roleCombos: Record<
-              string,
-              {
-                roles: {
-                  id: number;
-                  name: string;
-                  description: string | null;
-                }[];
-                matchCount: number;
-                wins: number;
-                losses: number;
-                winRate: number;
-                placements: Record<number, number>;
-              }
-            >;
-          }
-        >,
-      );
+      const { roleStats, comboRolesStats, playerStats } =
+        playerAndRolesAggregated(matches, result.roles);
 
       const duration = matches.reduce((acc, match) => {
         return acc + match.duration;
@@ -2038,73 +884,14 @@ export const gameRouter = createTRPCRouter({
         ownedBy: result.ownedBy,
         matches: matches,
         duration: duration,
-        players: Object.values(players).map((player) => ({
-          ...player,
-          coopWinRate:
-            player.coopMatches > 0 ? player.coopWins / player.coopMatches : 0,
-          competitiveWinRate:
-            player.competitiveMatches > 0
-              ? player.competitiveWins / player.competitiveMatches
-              : 0,
-          scoresheets: Object.values(player.scoresheets).map((scoresheet) => {
-            const sumScores = scoresheet.scores.reduce<number | null>(
-              (acc, score) => {
-                if (acc === null) return score.score;
-                if (score.score === null) return acc;
-                return acc + score.score;
-              },
-              null,
-            );
-            return {
-              ...scoresheet,
-              avgScore: sumScores ? sumScores / scoresheet.scores.length : null,
-              rounds: Object.values(scoresheet.rounds).map((round) => {
-                const sumScores = round.scores.reduce<number | null>(
-                  (acc, score) => {
-                    if (acc === null) return score.score;
-                    if (score.score === null) return acc;
-                    return acc + score.score;
-                  },
-                  null,
-                );
-                return {
-                  avgScore: sumScores ? sumScores / round.scores.length : null,
-                  ...round,
-                };
-              }),
-              winRate:
-                scoresheet.plays > 0 ? scoresheet.wins / scoresheet.plays : 0,
-            };
-          }),
-          roles: Object.values(player.roles).map((role) => ({
-            ...role,
-            winRate: role.wins / (role.wins + role.losses),
-          })),
-          roleCombos: Object.values(player.roleCombos).map((combo) => ({
-            ...combo,
-            winRate: combo.wins / (combo.wins + combo.losses),
-          })),
-        })),
+        players: playerStats,
         winRate: userWinRate,
         totalMatches: totalMatches,
         wonMatches: wonMatches,
         scoresheets: gameScoresheets,
         headToHead: headToHeadStats(matches),
-        roleStats: Object.values(roleStats).map((role) => ({
-          ...role,
-          winRate:
-            role.wins + role.losses > 0
-              ? role.wins / (role.wins + role.losses)
-              : 0,
-          players: Object.values(role.players),
-        })),
-        roleCombos: Object.values(comboRoles).map((combo) => ({
-          ...combo,
-          winRate:
-            combo.wins + combo.losses > 0
-              ? combo.wins / (combo.wins + combo.losses)
-              : 0,
-        })),
+        roleStats: roleStats,
+        roleCombos: comboRolesStats,
       };
     }),
   getGameName: protectedUserProcedure
